requirements.txt
streamlit==1.28.0
python-docx==1.1.0
python-pptx==0.6.23
PyYAML==6.0.1
Pillow==10.1.0
openpyxl==3.1.2
